requirements.txt
aiosqlite==0.17.0
asgiref==3.4.1
asyncpg==0.23.0
click==8.0.1
databases==0.4.3
fastapi==0.66.0
greenlet==1.1.0
h11==0.12.0
httptools==0.2.0
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.18.0
PyYAML==5.4.1
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
